fix: changed the name of the project
