+++ settings.gradle.kts
@@ -1,3 +1,3 @@
 rootProject.name = "trees-6"
 include("app")
-include("BSTrees")
+include("lib")
